refactor(metrics): toSortedMetrics を独立させる
